pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-usage-monitor"
version = "1.2.0"
description = "System tray app for monitoring Claude Code usage stats"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Andrew Barnes", email = "andrew.jaguars@gmail.com"},
]
keywords = ["claude", "claude-code", "usage", "monitor", "system-tray"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "pystray>=0.19",
    "Pillow>=10.0",
]

[project.scripts]
claude-usage = "claude_usage_monitor.app:main"

[project.urls]
Homepage = "https://github.com/Bortlesboat/claude-usage-monitor"
Repository = "https://github.com/Bortlesboat/claude-usage-monitor"
Issues = "https://github.com/Bortlesboat/claude-usage-monitor/issues"
